requirements.txt
fastapi
motor
python-jose[cryptography]
python-dotenv
openai
uvicorn
python-multipart